AIOOBE in BundleGenerator.ChunkGenerator.iterate when there are no outgoing changes on push

+++ src/org/tmatesoft/hg/core/HgPushCommand.java
@@ -87,13 +87,14 @@ public class HgPushCommand extends HgAbstractCommand<HgPushCommand> {
 			} else {
 				outgoing = new RevisionSet(l);
 			}
+			HgBundle b = null;
 			if (!outgoing.isEmpty()) {
 				//
 				// prepare bundle
 				BundleGenerator bg = new BundleGenerator(implRepo);
 				File bundleFile = bg.create(outgoing.asList());
 				progress.worked(20);
-				HgBundle b = new HgLookup(repo.getSessionContext()).loadBundle(bundleFile);
+				b = new HgLookup(repo.getSessionContext()).loadBundle(bundleFile);
 				//
 				// send changes
 				remoteRepo.unbundle(b, comparator.getRemoteHeads());
@@ -137,7 +138,9 @@ public class HgPushCommand extends HgAbstractCommand<HgPushCommand> {
 			}
 			// XXX WTF is obsolete in namespaces key??
 			progress.worked(5);
-			b.unlink(); // keep the file only in case of failure
+			if (b != null) {
+				b.unlink(); // keep the file only in case of failure
+			}
 		} catch (IOException ex) {
 			throw new HgIOException(ex.getMessage(), null); // XXX not a nice idea to throw IOException from BundleGenerator#create
 		} catch (HgRepositoryNotFoundException ex) {
